test_cubic_spline

+++ src/my_project/utility/Norm.hpp
@@ -21,8 +21,6 @@ enum NormType
  * @param normType тип нормы
  * @return norm соответствующая норма вектора
  */
-
-
 template <typename T, int Type>
 struct Norm{
     static double get_norm(const std::vector<T>& vector);
@@ -30,7 +28,7 @@ struct Norm{
 
 template<>
 double Norm<double, NormType::FirstNorm>::get_norm(const std::vector<double> &vector) {
-    double norm = static_cast<double>(0);
+    double norm = 0.;
     for (const auto& elm: vector)
     {
         double abs = std::abs(elm);
@@ -41,7 +39,7 @@ double Norm<double, NormType::FirstNorm>::get_norm(const std::vector<double> &ve
 
 template<>
 double Norm<double, NormType::SecondNorm>::get_norm(const std::vector<double> &vector) {
-    double norm = static_cast<double>(0);
+    double norm = 0.;
     for (const auto& elm: vector)
         norm += std::abs(elm);
     return norm;
@@ -54,35 +52,5 @@ double Norm<double, NormType::ThirdNorm>::get_norm(const std::vector<double> &ve
         norm += vector[i] * vector[i];
     return std::sqrt(norm);
 }
-//{
-//    T norm = static_cast<T>(0);
-//    if (normType == NormType::FirstNorm)
-//    {
-//        for (const auto& elm: vector)
-//        {
-//            T abs = std::abs(elm);
-//            if (abs > norm) norm = abs;
-//        }
-//    }
-//
-//    if (normType == NormType::SecondNorm)
-//    {
-//        for (const auto& elm: vector)
-//        {
-//            norm += std::abs(elm);
-//        }
-//    }
-//
-//    if (normType == NormType::ThirdNorm)
-//    {
-//        for(const auto& elm: vector)
-//        {
-//            norm += elm * elm;
-//        }
-//        norm = std::sqrt(norm);
-//    }
-//
-//    return norm;
-//}
 
 #endif//SLAE_NORM_HPP
